Refactor: add authCheck middleware change to receive position by parameter
